wire: sbe v8 — market budget rides in price, trim status ts, drop dead singles
Schema order-schema.xml v7 -> v8, wire-format trim driven by consumer-usage
evidence (no consumer read what was removed):

- CreateOrder: totalPrice removed; a MARKET buy's spend budget now rides in
  price (the two fields were never meaningful simultaneously). Block 46 -> 38B,
  ~ -8B per order in the cluster log. SbeDemuxer feeds the internal command's
  totalPrice from price-when-MARKET and keeps command price 0 for MARKET, so
  engine behaviour and status egress are byte-identical to v7.
- OrderStatusBatch orders group: per-entry timestamp removed (message-level
  timestamp stays). Entry block 75 -> 67B, -8B per status entry. No consumer
  read it: OMS's listener has no per-entry ts, gateway's onOrderStatusBatch is
  a deliberate no-op, loadtest reads only omsOrderId.
- Dead single-event messages TradeExecution (4), OrderStatusUpdate (5),
  OrderBookUpdate (6) deleted; zero encoder usage in match-cluster. Ids stay
  reserved, nothing renumbered. Loadtest's single-status decode branch removed
  (batch ack-coverage counting unchanged).
- ClusterHeartbeat.nodeId int64 -> uint8 (member ids are 0-2): 16 -> 9B, -7B
  per heartbeat. Gateway never decodes the body (template-id match only).
- TradeExecutionBatch, BookSnapshot, BookDelta, OpenOrdersSnapshot,
  RequestOpenOrdersSnapshot and both journal schemas untouched by decision
  (money keys + consumer evidence).

Compatibility: fields are REMOVED, so v8 decoders cannot read pre-v8 cluster
logs or streams. Accepted: zero customers; dev clusters restart from a fresh
log on deploy. Determinism goldens are behavioural event traces fed by
in-memory commands (ScenarioRunner never touches the wire), so they are
unaffected and unchanged — DeterminismCorpusTest passes against the committed
goldens. OMS (separate repo) must move the MARKET budget into price when it
adopts the v8 codecs.

+++ match-cluster/src/main/java/com/match/infrastructure/persistence/AppClusteredService.java
@@ -927,7 +927,9 @@ public class AppClusteredService implements ClusteredService {
     private void sendEgressKeepWarm() {
         keepWarmEncoder.wrapAndApplyHeader(keepWarmBuffer, 0, keepWarmHeaderEncoder);
         keepWarmEncoder
-            .nodeId(cluster.memberId())
+            // v8: nodeId is uint8 on the wire (member ids are 0-2); SBE's Java codec
+            // takes uint8 as short.
+            .nodeId((short) cluster.memberId())
             .timestamp(System.currentTimeMillis());
         int length = MessageHeaderEncoder.ENCODED_LENGTH + keepWarmEncoder.encodedLength();
